requirement.txt
supabase
python-dotenv
langchain_openai